test: refactor test to use fixtures
Use fixtures for setup that will reused in other tests
