+++ tests/test_gw/test_equivalent_proposal.py
@@ -10,17 +10,10 @@ import numpy as np
 import pytest
 
 from nessai import config
-from nessai.model import Model as BaseModel
 from nessai.gw.proposal import GWFlowProposal
 from nessai.gw.legacy import LegacyGWFlowProposal
 from nessai.flowsampler import FlowSampler
-from nessai.livepoint import dict_to_live_points
-
-try:
-    import bilby
-except ImportError:
-    print("Could not import bilby, some tests will not work")
-
+from nessai.utils.testing import assert_structured_arrays_equal
 
 logger = logging.getLogger(__name__)
 
@@ -42,28 +35,6 @@ def update_config():
     config.NON_SAMPLING_DEFAULTS = [0.0, 0.0, 0]
     yield
     config.NON_SAMPLING_DEFAULTS = original
-
-
-@pytest.fixture(scope="module")
-def injection_parameters():
-    injection_parameters = dict(
-        mass_1=36.0,
-        mass_2=29.0,
-        a_1=0.4,
-        a_2=0.3,
-        tilt_1=0.5,
-        tilt_2=1.0,
-        phi_12=1.7,
-        phi_jl=0.3,
-        luminosity_distance=2000.0,
-        theta_jn=0.4,
-        psi=2.659,
-        phase=1.3,
-        geocent_time=1126259642.413,
-        ra=1.375,
-        dec=-1.2108,
-    )
-    return injection_parameters
 
 
 @pytest.fixture(scope="function")
@@ -135,11 +106,15 @@ def kwargs(legacy_kwargs):
         ["chirp_mass", "geocent_time"],
         ["chirp_mass", "mass_ratio"],
         ["chirp_mass", "luminosity_distance"],
+        ["mass_ratio", "chirp_mass"],
+        ["mass_ratio", "ra", "dec"],
+        ["ra", "dec", "mass_ratio"],
     ],
 )
 @pytest.mark.parametrize("compute_radius", [False, True])
 @pytest.mark.integration_test
 def test_parameter(
+    get_bilby_gw_model,
     parameters,
     injection_parameters,
     kwargs,
@@ -154,118 +129,15 @@ def test_parameter(
     log-Jacobian determinants are also the same.
     """
     outdir = str(tmpdir.mkdir("test"))
-    priors = bilby.gw.prior.BBHPriorDict()
-    fixed_params = [
-        "chirp_mass",
-        "mass_ratio",
-        "phi_12",
-        "phi_jl",
-        "a_1",
-        "a_2",
-        "tilt_1",
-        "tilt_2",
-        "ra",
-        "dec",
-        "luminosity_distance",
-        "geocent_time",
-        "theta_jn",
-        "psi",
-    ]
-    try:
-        fixed_params.remove(parameters)
-    except ValueError:
-        for p in parameters:
-            fixed_params.remove(p)
-    priors["geocent_time"] = bilby.core.prior.Uniform(
-        minimum=injection_parameters["geocent_time"] - 0.1,
-        maximum=injection_parameters["geocent_time"] + 0.1,
-        name="geocent_time",
-        latex_label="$t_c$",
-        unit="$s$",
-    )
-    for key in fixed_params:
-        if key in injection_parameters:
-            priors[key] = injection_parameters[key]
-        else:
-            priors.pop(key)
 
-    priors["mass_1"] = injection_parameters["mass_1"]
-    priors["mass_2"] = injection_parameters["mass_2"]
-
-    waveform_generator = bilby.gw.WaveformGenerator(
-        duration=1,
-        sampling_frequency=256,
-        frequency_domain_source_model=bilby.gw.source.lal_binary_black_hole,
-    )
-
-    likelihood = bilby.gw.GravitationalWaveTransient(
-        interferometers=["H1"],
-        waveform_generator=waveform_generator,
-        priors=priors,
-        phase_marginalization=True,
-        distance_marginalization=False,
-        time_marginalization=False,
-        reference_frame="sky",
-    )
-
-    likelihood = bilby.core.likelihood.ZeroLikelihood(likelihood)
-
-    def log_prior(theta):
-        return priors.ln_prob(theta, axis=0)
-
-    search_parameter_keys = []
-    for key in priors:
-        if (
-            isinstance(priors[key], bilby.core.prior.Prior)
-            and priors[key].is_fixed is False
-        ):
-            search_parameter_keys.append(key)
-
-    def log_likelihood(theta):
-        params = {key: t for key, t in zip(search_parameter_keys, theta)}
-
-        likelihood.parameters.update(params)
-        return likelihood.log_likelihood_ratio()
-
-    class Model(BaseModel):
-        def __init__(self, names, priors):
-            self.names = names
-            self.priors = priors
-            self._update_bounds()
-
-        @staticmethod
-        def log_likelihood(x, **kwargs):
-            theta = [x[n].item() for n in search_parameter_keys]
-            return log_likelihood(theta)
-
-        @staticmethod
-        def log_prior(x, names=None, **kwargs):
-            if names is None:
-                names = search_parameter_keys
-            theta = {n: x[n] for n in names}
-            return log_prior(theta)
-
-        def _update_bounds(self):
-            self.bounds = {
-                key: [self.priors[key].minimum, self.priors[key].maximum]
-                for key in self.names
-            }
-
-        def new_point(self, N=1):
-            prior_samples = self.priors.sample(size=N)
-            samples = {n: prior_samples[n] for n in self.names}
-            self._update_bounds()
-            return dict_to_live_points(samples)
-
-        def new_point_log_prob(self, x):
-            return self.log_prior(x)
+    model = get_bilby_gw_model(parameters, injection_parameters)
 
     reparameterisations = kwargs.pop("reparameterisations")
     for k in list(reparameterisations.keys()):
         if k == "sky-ra-dec":
-            if any(p in fixed_params for p in ["ra", "dec"]):
+            if not all(p in model.names for p in ["ra", "dec"]):
                 del reparameterisations["sky-ra-dec"]
-        elif k not in search_parameter_keys:
+        elif k not in model.names:
             del reparameterisations[k]
 
     if "luminosity_distance" in parameters:
@@ -273,7 +145,6 @@ def test_parameter(
             "uniform_distance_parameter"
         ] = True
 
-    model = Model(search_parameter_keys, priors)
     sampler = FlowSampler(
         model, output=outdir, flow_class=LegacyGWFlowProposal, **legacy_kwargs
     )
@@ -323,7 +194,7 @@ def test_parameter(
         np.testing.assert_array_almost_equal_nulp(lj_re, lj_re_new)
         np.testing.assert_array_equal(log_p, log_p_new)
         try:
-            np.testing.assert_array_equal(x_prime, x_prime_new)
+            assert_structured_arrays_equal(x_prime, x_prime_new)
         except (AssertionError, TypeError):
             flag = {n: False for n in x_prime_new.dtype.names}
             for n in x_prime_new.dtype.names:
@@ -333,7 +204,6 @@ def test_parameter(
                 if np.allclose(x_prime[legacy_name], x_prime_new[n]):
                     logger.critical(f"{n} is equivalent to {legacy_name}")
                     flag[n] = True
-
             assert all(v for v in flag.values()), print(f"Flags: {flag}")
         test_results[test] = True
